refactor: app과 config 분리
