+++ app/src/main/java/com/joonhyeok/config/opentelemetry/OpenTelemetryConfig.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.common.config;
+package com.joonhyeok.config.opentelemetry;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.trace.Tracer;
