+++ app/src/main/java/com/joonhyeok/config/redis/RedisCacheConfig.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.common.config;
+package com.joonhyeok.config.redis;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
