+++ app/src/main/java/com/joonhyeok/config/redis/RedissonConfig.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.common.config;
+package com.joonhyeok.config.redis;
 
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
