+++ app/src/main/java/com/joonhyeok/config/web/HttpLoggingFilter.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.common.filter;
+package com.joonhyeok.config.web;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
